Update ApplozicSwift to 5.5.0
